Fixed a bug:
If clicking an entry in the search result list, the corresponding file will open in a tab. If placing the cursor at a word or selecting a word, searching either for all files or the current file, the initial word will not switch to the one selected, but remains the old one in the last time.